MK3-main/requirements.txt
wget
ffmpeg-python
motor
py-tgcalls==0.9.7
pyrofork
tgcrypto
youtube-search
youtube-search-python
yt_dlp
python-dotenv
dnspython
gitpython
pytube
asgiref
pyshorteners
aiofiles
aiohttp
requests
pillow==9.5.0
heroku3
psutil
future
wget
speedtest-cli
pyromod~=1.5
googletrans
telethon
redis
unidecode
numpy 
redis
python-arq
